Refactor and make the project cdylib
