+++ source/loaders/rs_loader/.gitignore
@@ -1,0 +1,3 @@
+/target
+Cargo.lock
+
