+++ source/loaders/rs_loader/Cargo.toml
@@ -1,10 +1,18 @@
 [package]
-name = "compiler"
+name = "rs_loader"
 version = "0.1.0"
 edition = "2018"
+
+[lib]
+name = "rs_loader"
+crate-type = ["cdylib"]      # Creates dynamic lib
+# crate-type = ["staticlib"] # Creates static lib
 
 [package.metadata.rust-analyzer]
 rustc_private = true
 
 [profile.release]
-opt-level="z"
+opt-level = "z"
+
+[workspace]
+members = ["functions_extractor"]
