+++ source/loaders/rs_loader/src/lifecycle/load_from_file.rs
@@ -1,0 +1,5 @@
+use crate::{PointerToCvoid};
+
+#[no_mangle]
+// There's nothing necessary to do at this step yet
+pub extern "C" fn rs_loader_impl_load_from_file(_loader_impl: PointerToCvoid) {}
